requirements.txt
cascade-ml>=0.12.0,<0.13
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4